requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0
langchain-tavily>=0.1.0
langgraph>=0.0.40
python-dotenv>=1.0.0
typing-extensions>=4.5.0
arxiv>=1.4.0
wikipedia>=1.4.0
pydantic>=2.5.0
gunicorn>=21.2.0
